GPKG: fix crash when opening an empty raster with USE_TILE_EXTENT=YES
git-svn-id: https://svn.osgeo.org/gdal/trunk@37524 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/gdrivers/gpkg.py
@@ -249,6 +249,15 @@ def gpkg_1():
         return 'fail'
     ds = None
 
+    # Test USE_TILE_EXTENT=YES with empty table
+    ds = gdal.OpenEx('tmp/tmp.gpkg', gdal.OF_UPDATE)
+    ds.ExecuteSQL('DELETE FROM tmp')
+    ds = None
+    ds = gdal.OpenEx('tmp/tmp.gpkg', gdal.OF_RASTER, open_options = ['USE_TILE_EXTENT=YES'])
+    if ds is not None:
+        gdaltest.post_reason('fail')
+        return 'fail'
+
     os.remove('tmp/tmp.gpkg')
 
     # Without padding
